macbot: fixed poll cmd input error

+++ macbot/commands/poll.js
@@ -104,11 +104,15 @@ module.exports = {
           for (var i = 0; i < input.length; i++) {
             embed.addField(input[i].slice(0, 3) + msg.reaction_count[i], input[i].slice(2), true)
           }
-          msg.reaction_count.unshift('piechart')
-          const image = canvas_cmd.execute(message, msg.reaction_count);
-          embed.setImage('attachment://canvas.png')
-          msg.delete();
-          message.channel.send({files: [image], embed: embed});
+          if (!msg.total_votes) {
+            msg.edit(embed);
+          } else {
+            msg.reaction_count.unshift('piechart')
+            const image = canvas_cmd.execute(message, msg.reaction_count);
+            embed.setImage('attachment://canvas.png')
+            msg.delete();
+            message.channel.send({files: [image], embed: embed});
+          }
         });
 
         collector.on('error', error => console.log('collector error'))
@@ -124,7 +128,7 @@ module.exports = {
       type = 2; // unlimited votes per person
     }
 
-    let length = 30;
+    let length = 120;
     if (!isNaN(args[0]) && args[0] <= 600) {
       length = args[0];
       args.shift();
@@ -134,7 +138,7 @@ module.exports = {
 
     const input = format_input(args);
 
-    if (input.length = 1) {
+    if (input.length === 1) {
       return message.channel.send('you need at least two options');
     }
